ml_backend/requirements.txt
pifastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.4.1
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
